Sketcher: Refactor DrawSketchHandlerLine and default widget support
===================================================================

This is a reimplementation of DrawSketchHandlerLine using the new framework.

It is documented to serve as an example of Method 1 of using DSHandlerDefaultWidget.

DrawSketchHandlerLine does not require additional data members with respect to the framework. Therefore,
it is more concise to just instantiate and specialise DSHandlerDefaultWidget.

It also shows how the default implementation based on the SelectModeT type may be leveraged to even reduce
further the amount of code necessary.